coding correction in classes
map: added walls to avoid the game from crashing for out of boundaries in map.
Collision detection and Tilemanager: Tile Array Name corrected ( it was mistyped)
Player: The character starts from the middle of the screen now.
GamePanel: The panel size was in fuction to the world size, it was modified to be a fix value.